Fix the declaration bug in converse/machinetest
Change double to int in declarations used in tests/converse/machinetest

+++ tests/converse/machinetest/multiping.C
@@ -24,7 +24,7 @@ enum { maxMsgSize = 1 << 18 };
 //Variable declarations
 CpvDeclare(int,msgSize);
 CpvDeclare(int,cycleNum);
-CpvDeclare(double, ackCount);
+CpvDeclare(int, ackCount);
 CpvDeclare(int, twoway);
 
 CpvDeclare(int,exitHandler);
