+++ tests/converse/machinetest/pingall.C
@@ -21,7 +21,7 @@ enum { maxMsgSize = 1 << 18 };
 CpvDeclare(int,msgSize);
 CpvDeclare(int,cycleNum);
 CpvDeclare(int,sizeNum);
-CpvDeclare(double, ackCount);
+CpvDeclare(int, ackCount);
 CpvDeclare(int, twoway);
 CpvDeclare(int,exitHandler);
 CpvDeclare(int,node0Handler);
